# api/openapi/service.yaml
openapi: 3.0.3
info:
  title: billing-service
  version: 0.1.0
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: https://billing-service.internal.gonkagate.local
    description: placeholder internal service base URL
security: []
...
paths:
...
  /internal/billing/v1/accounts/resolve:
    post:
      tags: [billing]
      summary: Resolve represented caller identity to a billing account
      operationId: resolveBillingAccount
      security:
        - ServiceBearerAuth: []
      x-route-scopes: [billing.accounts.resolve]
      x-security-decision:
        exposure: protected
        rationale: service-to-service account binding and migration-readiness path
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AccountResolveRequest"
      responses:
        "200":
          description: account resolve outcome
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AccountResolveResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "423":
          $ref: "#/components/responses/Locked"
        "503":
          $ref: "#/components/responses/ServiceUnavailable"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /internal/billing/v1/accounts/{accountScopeKey}/balance:
    get:
      tags: [billing]
      summary: Read authoritative billing balance and active exposure
      operationId: readBillingBalance
      security:
        - ServiceBearerAuth: []
      x-route-scopes: [billing.balances.read]
      x-security-decision:
        exposure: protected
        rationale: service-to-service migrated balance readback path
      parameters:
        - name: accountScopeKey
          in: path
          required: true
          schema:
            type: string
            minLength: 1
        - name: representedSubjectId
          in: query
          required: true
          schema:
            type: string
            minLength: 1
        - name: traceRequestId
          in: query
          required: true
          schema:
            type: string
            minLength: 1
        - name: deadlineAtEpochMs
          in: query
          required: true
          schema:
            type: integer
            format: int64
            minimum: 0
      responses:
        "200":
          description: balance read outcome
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BalanceReadResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "423":
          $ref: "#/components/responses/Locked"
        "503":
          $ref: "#/components/responses/ServiceUnavailable"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /internal/billing/v1/usage/reservations:
    post:
      tags: [billing]
      summary: Create or read back migrated usage reserve lineage
      operationId: reserveUsage
      security:
        - ServiceBearerAuth: []
      x-route-scopes: [billing.usage.write]
      x-security-decision:
        exposure: protected
        rationale: money-lineage write path for usage reserve under billing authority
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UsageReserveRequest"
      responses:
        "200":
          description: usage reserve outcome
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UsageOperationResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "409":
          $ref: "#/components/responses/Conflict"
        "422":
          $ref: "#/components/responses/UnprocessableEntity"
        "423":
          $ref: "#/components/responses/Locked"
        "429":
          $ref: "#/components/responses/TooManyRequests"
        "503":
          $ref: "#/components/responses/ServiceUnavailable"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /internal/billing/v1/usage/finalizations:
    post:
      tags: [billing]
      summary: Finalize a usage operation within microlease authority
      operationId: finalizeUsage
      security:
        - ServiceBearerAuth: []
      x-route-scopes: [billing.usage.write]
      x-security-decision:
        exposure: protected
        rationale: money-effect terminal write path for usage finalization
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UsageTerminalRequest"
      responses:
        "200":
          description: usage finalization outcome
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UsageOperationResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "409":
          $ref: "#/components/responses/Conflict"
        "422":
          $ref: "#/components/responses/UnprocessableEntity"
        "423":
          $ref: "#/components/responses/Locked"
        "503":
          $ref: "#/components/responses/ServiceUnavailable"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /internal/billing/v1/usage/write-offs:
    post:
      tags: [billing]
      summary: Write off or release a usage operation explicitly
      operationId: writeOffUsage
      security:
        - ServiceBearerAuth: []
      x-route-scopes: [billing.usage.write]
      x-security-decision:
        exposure: protected
        rationale: explicit support-safe write-off path for unchargeable usage
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UsageTerminalRequest"
      responses:
        "200":
          description: usage write-off outcome
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UsageOperationResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "409":
          $ref: "#/components/responses/Conflict"
        "422":
          $ref: "#/components/responses/UnprocessableEntity"
        "423":
          $ref: "#/components/responses/Locked"
        "503":
          $ref: "#/components/responses/ServiceUnavailable"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /internal/billing/v1/usage/reversals:
    post:
      tags: [billing]
      summary: Reverse a previously committed usage effect
      operationId: reverseUsage
      security:
        - ServiceBearerAuth: []
      x-route-scopes: [billing.usage.write]
      x-security-decision:
        exposure: protected
        rationale: explicit compensation path for committed usage effects
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UsageReversalRequest"
      responses:
        "200":
          description: usage reversal outcome
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UsageOperationResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "409":
          $ref: "#/components/responses/Conflict"
        "422":
          $ref: "#/components/responses/UnprocessableEntity"
        "423":
          $ref: "#/components/responses/Locked"
        "503":
          $ref: "#/components/responses/ServiceUnavailable"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /internal/billing/v1/usage/readback:
    post:
      tags: [billing]
      summary: Read usage operation state by durable usage identity
      operationId: readUsageOperation
      security:
        - ServiceBearerAuth: []
      x-route-scopes: [billing.usage.read]
      x-security-decision:
        exposure: protected
        rationale: account-bound usage readback path for ambiguous outcomes
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UsageReadbackRequest"
      responses:
        "200":
          description: usage readback outcome
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UsageOperationReadbackResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "503":
          $ref: "#/components/responses/ServiceUnavailable"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /internal/billing/v1/microleases/issue:
    post:
      tags: [billing]
      summary: Issue or replenish a billing microlease
      operationId: issueMicrolease
      security:
        - ServiceBearerAuth: []
      x-route-scopes: [billing.microleases.write]
      x-security-decision:
        exposure: protected
        rationale: money-affecting service-to-service microlease issuance path
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MicroleaseIssueRequest"
      responses:
        "200":
          description: issue or replay outcome
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MicroleaseIssueResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "409":
          $ref: "#/components/responses/Conflict"
        "422":
          $ref: "#/components/responses/UnprocessableEntity"
        "423":
          $ref: "#/components/responses/Locked"
        "429":
          $ref: "#/components/responses/TooManyRequests"
        "503":
          $ref: "#/components/responses/ServiceUnavailable"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /internal/billing/v1/microleases/readback:
    post:
      tags: [billing]
      summary: Read a microlease or stored operation outcome
      operationId: readMicrolease
      security:
        - ServiceBearerAuth: []
      x-route-scopes: [billing.microleases.read]
      x-security-decision:
        exposure: protected
        rationale: support-safe service-to-service microlease readback path
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MicroleaseReadbackRequest"
      responses:
        "200":
          description: readback outcome
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MicroleaseReadbackResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "503":
          $ref: "#/components/responses/ServiceUnavailable"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /internal/billing/v1/microleases/close:
    post:
      tags: [billing]
      summary: Close or checkpoint a microlease with allocator proof
      operationId: closeMicrolease
      security:
        - ServiceBearerAuth: []
      x-route-scopes: [billing.microleases.write]
      x-security-decision:
        exposure: protected
        rationale: money-affecting service-to-service close and release path
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MicroleaseCloseRequest"
      responses:
        "200":
          description: close or checkpoint outcome
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MicroleaseCloseResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "409":
          $ref: "#/components/responses/Conflict"
        "422":
          $ref: "#/components/responses/UnprocessableEntity"
        "423":
          $ref: "#/components/responses/Locked"
        "503":
          $ref: "#/components/responses/ServiceUnavailable"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /internal/billing/v1/operations/readback:
    post:
      tags: [billing]
      summary: Read a stored operation outcome
      operationId: readBillingOperation
      security:
        - ServiceBearerAuth: []
      x-route-scopes: [billing.operations.read]
      x-security-decision:
        exposure: protected
        rationale: support-safe service-to-service operation readback path
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BillingOperationReadbackRequest"
      responses:
        "200":
          description: operation readback outcome
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BillingOperationReadbackResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "503":
          $ref: "#/components/responses/ServiceUnavailable"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /internal/billing/v1/reconciliation/cases:
    get:
      tags: [billing]
      summary: List support-safe reconciliation cases
      operationId: listReconciliationCases
      security:
        - ServiceBearerAuth: []
      x-route-scopes: [billing.reconciliation.read]
      x-security-decision:
        exposure: protected
        rationale: scoped support readback for reconciliation cases
      parameters:
        - name: accountScopeKey
          in: query
          required: false
          schema:
            type: string
            minLength: 1
        - name: state
          in: query
          required: false
          schema:
            type: string
            enum: [open, leased, waiting_evidence, manual_review, resolved, canceled]
        - name: severity
          in: query
          required: false
          schema:
            type: string
            enum: [low, medium, high, critical]
      responses:
        "200":
          description: reconciliation cases
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReconciliationCasesResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "503":
          $ref: "#/components/responses/ServiceUnavailable"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /internal/billing/v1/admin/accounts/{accountScopeKey}/ledger:
    get:
      tags: [billing]
      summary: Read support-safe ledger history for an account
      operationId: readAdminLedger
      security:
        - ServiceBearerAuth: []
      x-route-scopes: [billing.admin.read]
      x-security-decision:
        exposure: protected
        rationale: scoped internal ledger readback without mutation authority
      parameters:
        - name: accountScopeKey
          in: path
          required: true
          schema:
            type: string
            minLength: 1
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 50
      responses:
        "200":
          description: ledger readback
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminLedgerResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "503":
          $ref: "#/components/responses/ServiceUnavailable"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /internal/billing/v1/admin/accounts/{accountScopeKey}/exposure:
    get:
      tags: [billing]
      summary: Read active exposure and worker gate state for an account
      operationId: readAdminExposure
      security:
        - ServiceBearerAuth: []
      x-route-scopes: [billing.admin.read]
      x-security-decision:
        exposure: protected
        rationale: scoped internal exposure readback without mutation authority
      parameters:
        - name: accountScopeKey
          in: path
          required: true
          schema:
            type: string
            minLength: 1
      responses:
        "200":
          description: exposure readback
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminExposureResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "503":
          $ref: "#/components/responses/ServiceUnavailable"
        "500":
          $ref: "#/components/responses/InternalServerError"
components:
  securitySchemes:
    ServiceBearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Platform-issued service identity token for protected internal billing routes.
  schemas:
...
    ContractVersion:
      type: string
      enum: [v1]
...
    AccountState:
      type: string
      enum: [active, suspended, manual_review, closed, not_found, not_ready]
...
    AccountResolveRequest:
      type: object
      required:
        - routeContractId
        - contractVersion
        - callerContext
        - representedUserContext
        - deadlineAtEpochMs
        - traceRequestId
      additionalProperties: false
      properties:
        routeContractId:
          type: string
          enum: [billing.accounts.resolve]
        contractVersion:
          $ref: "#/components/schemas/ContractVersion"
        callerContext:
          $ref: "#/components/schemas/CallerContext"
        representedUserContext:
          $ref: "#/components/schemas/RepresentedUserContext"
        deadlineAtEpochMs:
          type: integer
          format: int64
          minimum: 0
        traceRequestId:
          type: string
          minLength: 1
    AccountResolveResponse:
      type: object
      required:
        - resultCode
        - accountScopeKey
        - accountState
        - importState
        - migrationState
        - balanceReadEligible
        - retryable
      additionalProperties: false
      properties:
        resultCode:
          $ref: "#/components/schemas/AccountResolveResultCode"
        accountScopeKey:
          type: string
          minLength: 1
        billingAccountId:
          type: string
          minLength: 1
        accountState:
          $ref: "#/components/schemas/AccountState"
        importState:
          $ref: "#/components/schemas/ImportState"
        migrationState:
          type: string
          enum: [legacy, shadow, internal_cohort, migrated, rollback]
        balanceReadEligible:
          type: boolean
        retryable:
          type: boolean
        failureClass:
          type: string
          minLength: 1
        balanceVersion:
          type: integer
          format: int64
          minimum: 0
...
  responses:
    BadRequest:
      description: request is malformed or invalid
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
...